Made some implementation changes

+++ 6.1 Auto Mapping Objects - Exercise/Game-Store/src/main/java/com/example/gamestore/CommandLineRunnerImpl.java
@@ -41,7 +41,7 @@ public class CommandLineRunnerImpl implements CommandLineRunner {
                 case "Logout" -> userService
                         .logout();
                 case "AddGame" -> gameService
-                        .addGame(new GameAddDto(commands[1], new BigDecimal(commands[2]), Integer.parseInt(commands[3]),
+                        .addGame(new GameAddDto(commands[1], new BigDecimal(commands[2]), Double.parseDouble(commands[3]),
                                 commands[4], commands[5], commands[6], commands[7]));
             }
         }
